@@ -5,25 +5,37 @@ description = "Yet Another Picture Search Nonebot Plugin"
 authors = ["NekoAria"]
 license = "GPL-3.0-only"
 readme = "README.md"
-packages = [
-    { include = "YetAnotherPicSearch" },
-]
+packages = [{ include = "YetAnotherPicSearch" }]
 homepage = "https://github.com/NekoAria/YetAnotherPicSearch"
 repository = "https://github.com/NekoAria/YetAnotherPicSearch"
-keywords = ["nonebot", "ascii2d", "baidu", "e-hentai", "google", "iqdb", "saucenao", "tracemoe", "yandex", "anime", "danbooru", "doujin", "pixiv"]
+keywords = [
+    "nonebot",
+    "ascii2d",
+    "baidu",
+    "e-hentai",
+    "google",
+    "iqdb",
+    "saucenao",
+    "tracemoe",
+    "yandex",
+    "anime",
+    "danbooru",
+    "doujin",
+    "pixiv",
+]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = ">=3.9,<4"
 arrow = "^1.3.0"
 cachetools = "^5.3.2"
 httpx = "^0.25.2"
 lxml = "^4.9.3"
-nonebot-adapter-onebot = "^2.3.1"
-nonebot2 = "^2.1.2"
+nonebot-adapter-onebot = "^2.4.1"
+nonebot2 = "^2.2.0"
 PicImageSearch = "^3.9.7"
-pydantic = "^1.10.13"
+pydantic = "^2.6.4"
 pyquery = "^2.0.0"
-python-socks = {extras = ["asyncio"], version = "^2.4.4", optional = true}
+python-socks = { extras = ["asyncio"], version = "^2.4.4", optional = true }
 socksio = { version = "^1.0.0", optional = true }
 shelved-cache = "^0.3.1"
 tenacity = "^8.2.3"
@@ -32,12 +44,11 @@ tenacity = "^8.2.3"
 socks = ["python-socks", "socksio"]
 
 [tool.poetry.dev-dependencies]
+pre-commit = "^3.6.2"
 
 [tool.mypy]
 python_version = "3.8"
-plugins = [
-    "pydantic.mypy"
-]
+plugins = ["pydantic.mypy"]
 ignore_missing_imports = true
 implicit_reexport = true
 pretty = true
